Bridge swift and C
Synopsis: Get a minimal amount of C code to be called from a minimal
amount of swift code. It was every bit as difficult as it could have
been because of how many different domains are involved, some of which
are moving targets.

Minor:

Changed the pseudo-field name of fragments from “header” to “heading”
from two reasons:

  - “header” is an existing term from documents (markdown and html)
    and we don’t want to over-suggest a surface expression at this level

  - “heading” suggests where a (e.g.) ship is going, so it seems to
    suggest “where the fragment is headed”; i.e sort of like an
    objective or an aspirational goal for the fragment.

- instead of learning about swift types, we’re changing our C type
  from `int` to `long`.

(times)

  05-15 02:14  amass several blogs (with varying degrees of quality and
               currency) and contemplate which ones(s) to go deep on.
        03:33  begin reading tutorial at hackingswift.com
        11:42  begin again, reading that blog.
  05-17 02:47  try again, reading blog.
        04:23  system packages are deprecated.
        10:07  got it working. cried.
  05-20 01:44  refine your one proof of concept
        03:50  done. begin cleaning up commit message etc
        04:11  now

+++ pho/main.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 int main(int argc, char **argv){
-  int num = factorial(3);
+  int num = (int)factorial(3);  // meh
   printf("hello!!!! %d\n", num);
   return 0;
 }
