+++ pho/pho-bridge.c
@@ -1,6 +1,6 @@
 #include "pho-bridge.h"
 
-int factorial(int n) {
+long factorial(long n) {
   if (n == 0 || n == 1) return 1;
   return n * factorial(n-1);
 }
